+++ Setup-Your-Mac-via-Dialog.bash
@@ -19,7 +19,7 @@
 #   - Updates to `README.md`, `CONTRIBUTORS.md` and `CONTRIBUTING.md` [Pull Request No. 128](https://github.com/setup-your-mac/Setup-Your-Mac/pull/128); thanks, @robjschroeder!
 #   - Refactored the way `brandingBanner` variable is checked [Pull Request No. 131](https://github.com/setup-your-mac/Setup-Your-Mac/pull/131); thanks, @drtaru!
 #   - Increased minimum required version of swiftDialog to 2.4.0.4748
-#   - (Started code to) Leverage the new `subtitle` in `listitem` (using `progresstext`, but probably needs a dedicated `subtitle` field in `policyJSON`)
+#   - Leveraged the new `listitem: subtitle` option with a dedicated `subtitle` field in `policyJSON`
 #
 ####################################################################################################
 
@@ -35,7 +35,7 @@
 # Script Version and Jamf Pro Script Parameters
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-scriptVersion="1.14.0-b2"
+scriptVersion="1.14.0-b3"
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin
 scriptLog="${4:-"/var/log/org.churchofjesuschrist.log"}"                        # Parameter 4: Script Log Location [ /var/log/org.churchofjesuschrist.log ] (i.e., Your organization's default location for client-side logs)
 debugMode="${5:-"verbose"}"                                                     # Parameter 5: Debug Mode [ verbose (default) | true | false ]
@@ -937,6 +937,7 @@ function policyJSONConfiguration() {
                 "steps": [
                     {
                         "listitem": "Rosetta",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_8bac19160fabb0c8e7bac97b37b51d2ac8f38b7100b6357642d9505645d37b52",
                         "progresstext": "Rosetta enables a Mac with Apple silicon to use apps built for a Mac with an Intel processor.",
                         "trigger_list": [
@@ -952,6 +953,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "FileVault Disk Encryption",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_f9ba35bd55488783456d64ec73372f029560531ca10dfa0e8154a46d7732b913",
                         "progresstext": "FileVault is built-in to macOS and provides full-disk encryption to help prevent unauthorized access to your Mac.",
                         "trigger_list": [
@@ -963,6 +965,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Sophos Endpoint",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_c70f1acf8c96b99568fec83e165d2a534d111b0510fb561a283d32aa5b01c60c",
                         "progresstext": "You’ll enjoy next-gen protection with Sophos Endpoint which doesn’t rely on signatures to catch malware.",
                         "trigger_list": [
@@ -974,6 +977,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Sophos Endpoint Services (Remote)",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_0f68be689684a00a3a054d71a31e43e2362f96c16efa5a560fb61bc1bf41901c",
                         "progresstext": "Remotely validating Sophos Endpoint services …",
                         "trigger_list": [
@@ -985,6 +989,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Palo Alto GlobalProtect",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_acbf39d8904ad1a772cf71c45d93e373626d379a24f8b1283b88134880acb8ef",
                         "progresstext": "Use Palo Alto GlobalProtect to establish a Virtual Private Network (VPN) connection to Church headquarters.",
                         "trigger_list": [
@@ -996,6 +1001,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Palo Alto GlobalProtect Services (Remote)",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_709e8bdf0019e8faf9df85ec0a68545bfdb8bfa1227ac9bed9bba40a1fa8ff42",
                         "progresstext": "Remotely validating Palo Alto GlobalProtect services …",
                         "trigger_list": [
@@ -1007,6 +1013,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Final Configuration",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_4723e3e341a7e11e6881e418cf91b157fcc081bdb8948697750e5da3562df728",
                         "progresstext": "Finalizing Configuration …",
                         "trigger_list": [
@@ -1022,6 +1029,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Computer Inventory",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_ff2147a6c09f5ef73d1c4406d00346811a9c64c0b6b7f36eb52fcb44943d26f9",
                         "progresstext": "A listing of your Mac’s apps and settings — its inventory — is sent automatically to the Jamf Pro server daily.",
                         "trigger_list": [
@@ -1044,6 +1052,7 @@ function policyJSONConfiguration() {
                 "steps": [
                     {
                         "listitem": "Rosetta",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_8bac19160fabb0c8e7bac97b37b51d2ac8f38b7100b6357642d9505645d37b52",
                         "progresstext": "Rosetta enables a Mac with Apple silicon to use apps built for a Mac with an Intel processor.",
                         "trigger_list": [
@@ -1059,6 +1068,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "FileVault Disk Encryption",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_f9ba35bd55488783456d64ec73372f029560531ca10dfa0e8154a46d7732b913",
                         "progresstext": "FileVault is built-in to macOS and provides full-disk encryption to help prevent unauthorized access to your Mac.",
                         "trigger_list": [
@@ -1070,6 +1080,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Sophos Endpoint",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_c70f1acf8c96b99568fec83e165d2a534d111b0510fb561a283d32aa5b01c60c",
                         "progresstext": "You’ll enjoy next-gen protection with Sophos Endpoint which doesn’t rely on signatures to catch malware.",
                         "trigger_list": [
@@ -1081,6 +1092,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Sophos Endpoint Services (Local)",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_0f68be689684a00a3a054d71a31e43e2362f96c16efa5a560fb61bc1bf41901c",
                         "progresstext": "Locally validating Sophos Endpoint services …",
                         "trigger_list": [
@@ -1092,6 +1104,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Palo Alto GlobalProtect",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_acbf39d8904ad1a772cf71c45d93e373626d379a24f8b1283b88134880acb8ef",
                         "progresstext": "Use Palo Alto GlobalProtect to establish a Virtual Private Network (VPN) connection to Church headquarters.",
                         "trigger_list": [
@@ -1103,6 +1116,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Palo Alto GlobalProtect Services (Local)",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_709e8bdf0019e8faf9df85ec0a68545bfdb8bfa1227ac9bed9bba40a1fa8ff42",
                         "progresstext": "Locally validating Palo Alto GlobalProtect services …",
                         "trigger_list": [
@@ -1114,6 +1128,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Microsoft 365",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_1801d1fdd81e19ce5eb0e567371377e7995bff32947adb7a94c5feea760edcb5",
                         "progresstext": "Office is now Microsoft 365. Create, share, and collaborate with your favorite apps — all in one place — with Microsoft 365.",
                         "trigger_list": [
@@ -1129,6 +1144,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Microsoft Teams",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_dcb65709dba6cffa90a5eeaa54cb548d5ecc3b051f39feadd39e02744f37c19e",
                         "progresstext": "Microsoft Teams is a hub for teamwork in Office 365. Keep all your team’s chats, meetings and files together in one place.",
                         "trigger_list": [
@@ -1140,6 +1156,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Final Configuration",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_4723e3e341a7e11e6881e418cf91b157fcc081bdb8948697750e5da3562df728",
                         "progresstext": "Finalizing Configuration …",
                         "trigger_list": [
@@ -1155,6 +1172,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Computer Inventory",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_ff2147a6c09f5ef73d1c4406d00346811a9c64c0b6b7f36eb52fcb44943d26f9",
                         "progresstext": "A listing of your Mac’s apps and settings — its inventory — is sent automatically to the Jamf Pro server daily.",
                         "trigger_list": [
@@ -1177,6 +1195,7 @@ function policyJSONConfiguration() {
                 "steps": [
                     {
                         "listitem": "Rosetta",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_8bac19160fabb0c8e7bac97b37b51d2ac8f38b7100b6357642d9505645d37b52",
                         "progresstext": "Rosetta enables a Mac with Apple silicon to use apps built for a Mac with an Intel processor.",
                         "trigger_list": [
@@ -1192,6 +1211,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "FileVault Disk Encryption",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_f9ba35bd55488783456d64ec73372f029560531ca10dfa0e8154a46d7732b913",
                         "progresstext": "FileVault is built-in to macOS and provides full-disk encryption to help prevent unauthorized access to your Mac.",
                         "trigger_list": [
@@ -1203,6 +1223,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Sophos Endpoint",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_c70f1acf8c96b99568fec83e165d2a534d111b0510fb561a283d32aa5b01c60c",
                         "progresstext": "You’ll enjoy next-gen protection with Sophos Endpoint which doesn’t rely on signatures to catch malware.",
                         "trigger_list": [
@@ -1214,6 +1235,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Sophos Endpoint Services (Local)",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_0f68be689684a00a3a054d71a31e43e2362f96c16efa5a560fb61bc1bf41901c",
                         "progresstext": "Locally validating Sophos Endpoint services …",
                         "trigger_list": [
@@ -1225,6 +1247,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Sophos Endpoint Services (Remote)",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_0f68be689684a00a3a054d71a31e43e2362f96c16efa5a560fb61bc1bf41901c",
                         "progresstext": "Remotely validating Sophos Endpoint services …",
                         "trigger_list": [
@@ -1236,6 +1259,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Palo Alto GlobalProtect",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_acbf39d8904ad1a772cf71c45d93e373626d379a24f8b1283b88134880acb8ef",
                         "progresstext": "Use Palo Alto GlobalProtect to establish a Virtual Private Network (VPN) connection to Church headquarters.",
                         "trigger_list": [
@@ -1247,6 +1271,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Palo Alto GlobalProtect Services (Local)",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_709e8bdf0019e8faf9df85ec0a68545bfdb8bfa1227ac9bed9bba40a1fa8ff42",
                         "progresstext": "Locally validating Palo Alto GlobalProtect services …",
                         "trigger_list": [
@@ -1258,6 +1283,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Palo Alto GlobalProtect Services (Remote)",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_709e8bdf0019e8faf9df85ec0a68545bfdb8bfa1227ac9bed9bba40a1fa8ff42",
                         "progresstext": "Remotely validating Palo Alto GlobalProtect services …",
                         "trigger_list": [
@@ -1269,6 +1295,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Microsoft 365",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_1801d1fdd81e19ce5eb0e567371377e7995bff32947adb7a94c5feea760edcb5",
                         "progresstext": "Office is now Microsoft 365. Create, share, and collaborate with your favorite apps — all in one place — with Microsoft 365.",
                         "trigger_list": [
@@ -1284,6 +1311,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Microsoft Teams",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_dcb65709dba6cffa90a5eeaa54cb548d5ecc3b051f39feadd39e02744f37c19e",
                         "progresstext": "Microsoft Teams is a hub for teamwork in Office 365. Keep all your team’s chats, meetings and files together in one place.",
                         "trigger_list": [
@@ -1295,6 +1323,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Adobe Acrobat Reader",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_988b669ca27eab93a9bcd53bb7e2873fb98be4eaa95ae8974c14d611bea1d95f",
                         "progresstext": "Views, prints, and comments on PDF documents, and connects to Adobe Document Cloud.",
                         "trigger_list": [
@@ -1306,6 +1335,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Google Chrome",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_12d3d198f40ab2ac237cff3b5cb05b09f7f26966d6dffba780e4d4e5325cc701",
                         "progresstext": "Google Chrome is a browser that combines a minimal design with sophisticated technology to make the Web faster.",
                         "trigger_list": [
@@ -1317,6 +1347,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Final Configuration",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_4723e3e341a7e11e6881e418cf91b157fcc081bdb8948697750e5da3562df728",
                         "progresstext": "Finalizing Configuration …",
                         "trigger_list": [
@@ -1332,6 +1363,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Computer Inventory",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_ff2147a6c09f5ef73d1c4406d00346811a9c64c0b6b7f36eb52fcb44943d26f9",
                         "progresstext": "A listing of your Mac’s apps and settings — its inventory — is sent automatically to the Jamf Pro server daily.",
                         "trigger_list": [
@@ -1354,6 +1386,7 @@ function policyJSONConfiguration() {
                 "steps": [
                     {
                         "listitem": "Rosetta",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_8bac19160fabb0c8e7bac97b37b51d2ac8f38b7100b6357642d9505645d37b52",
                         "progresstext": "Rosetta enables a Mac with Apple silicon to use apps built for a Mac with an Intel processor.",
                         "trigger_list": [
@@ -1369,6 +1402,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "FileVault Disk Encryption",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_f9ba35bd55488783456d64ec73372f029560531ca10dfa0e8154a46d7732b913",
                         "progresstext": "FileVault is built-in to macOS and provides full-disk encryption to help prevent unauthorized access to your Mac.",
                         "trigger_list": [
@@ -1380,6 +1414,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Sophos Endpoint",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_c70f1acf8c96b99568fec83e165d2a534d111b0510fb561a283d32aa5b01c60c",
                         "progresstext": "You’ll enjoy next-gen protection with Sophos Endpoint which doesn’t rely on signatures to catch malware.",
                         "trigger_list": [
@@ -1395,6 +1430,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Palo Alto GlobalProtect",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_acbf39d8904ad1a772cf71c45d93e373626d379a24f8b1283b88134880acb8ef",
                         "progresstext": "Use Palo Alto GlobalProtect to establish a Virtual Private Network (VPN) connection to Church headquarters.",
                         "trigger_list": [
@@ -1410,6 +1446,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Final Configuration",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_4723e3e341a7e11e6881e418cf91b157fcc081bdb8948697750e5da3562df728",
                         "progresstext": "Finalizing Configuration …",
                         "trigger_list": [
@@ -1425,6 +1462,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Computer Inventory",
+                        "subtitle": "Subtitle goes here",
                         "icon": "https://ics.services.jamfcloud.com/icon/hash_ff2147a6c09f5ef73d1c4406d00346811a9c64c0b6b7f36eb52fcb44943d26f9",
                         "progresstext": "A listing of your Mac’s apps and settings — its inventory — is sent automatically to the Jamf Pro server daily.",
                         "trigger_list": [
@@ -3096,7 +3134,7 @@ for (( i=0; i<dialog_step_length; i++ )); do
     list_item_array+=("$listitem")
     icon=$(get_json_value "${policyJSON}" "steps[$i].icon")
     icon_url_array+=("$icon")
-    subtitle=$(get_json_value "${policyJSON}" "steps[$i].progresstext")
+    subtitle=$(get_json_value "${policyJSON}" "steps[$i].subtitle")
     subtitle_array+=("$subtitle")
 done